+++ packages/guide/src/app/core/types/user-data.type.ts
@@ -1,10 +1,10 @@
 import { ToDo } from "./to-do.type";
-import { SpecificDate } from "@ci/data-types";
 import { Checklist } from "../interfaces/checklist.interface";
+import { DashboardFilterData } from "../../my-coral-guide/types/dashboard-filter.type";
 
 export type UserData = {
     name: string;
-    myGuideDate: SpecificDate;
+    myGuideFilter: DashboardFilterData;
     todos: ToDo[];
     checklists: Record<string, Checklist>;
 }