+++ packages/guide/src/CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Changed
 
 - Fixed fish with date ranges not being shown properly in dashboard.
+- Fixed date/configuration for dashboard (My Guide) not being persisted.
 
 ### 0.21.0 2024-07-23